Some initial code cleanup
